feat: 보유 잔액이 부족하면 alert 메시지를 보여준다

+++ src/js/controller/ProductPurchase.js
@@ -3,6 +3,7 @@ import { SECTION_CONTAINER } from '../constants/constants.js';
 import PurchaseAmountModel from '../models/PurchaseAmount.ts';
 import ReturnedCoinModel from '../models/ReturnedCoin.ts';
 import ProductPurchaseView from '../views/ProductPurchaseView.js';
+import { validAffordablePrice } from '../utils/validation.js';
 
 export default class ProductPurchase {
   constructor(product, coin) {
@@ -13,8 +14,7 @@ export default class ProductPurchase {
     this.productPurchaseView = new ProductPurchaseView();
 
     on(SECTION_CONTAINER, '@purchase', this.#handlePurchaseAmount.bind(this));
-    on(SECTION_CONTAINER, '@quantity', this.#modifyProductQuantity.bind(this));
-    on(SECTION_CONTAINER, '@soldOut', this.#deleteSoldOutProduct.bind(this));
+    on(SECTION_CONTAINER, '@buy', this.#handleProductPurchase.bind(this));
   }
 
   initPurchase() {
@@ -34,18 +34,21 @@ export default class ProductPurchase {
     this.productPurchaseView.resetAmountInput();
   }
 
-  #modifyProductQuantity(e) {
+  #handleProductPurchase(e) {
+    const { index, product } = e.detail;
     try {
-      const { index, product } = e.detail;
-      this.productModel.modifyProduct(index, product);
+      validAffordablePrice(this.purchaseAmountModel.getAmount(), product.price);
+      if (product.quantity === 0) {
+        this.productModel.deleteProduct(index);
+      } else {
+        this.productModel.modifyProduct(index, product);
+      }
+
+      this.purchaseAmountModel.deductAmount(product.price);
+      this.productPurchaseView.renderTotalAmount(this.purchaseAmountModel.getAmount());
+      this.productPurchaseView.renderProducts(this.productModel.getProducts());
     } catch (error) {
       alert(error.message);
     }
-  }
-
-  #deleteSoldOutProduct(e) {
-    const { index } = e.detail;
-    this.productModel.deleteProduct(index);
-    this.productPurchaseView.renderProducts(this.productModel.getProducts());
   }
 }
